Fixed wrongly defined prototype method

+++ src/arduino_code/PIC16F184XX_pic_programmer.h
@@ -82,6 +82,6 @@ private:
 
 	// --------------- LOAD ADDRESS COMMAND --------------- //
 
-	void commandLoadPCAddress(unsigned long addr);
+	void commandLoadPCAddress(unsigned int addr);
 
 };
